use make for dependency resolution
One of make's strengths is its ability to manage dependencies. By creating
recipes for each file to be created, installed, or generated, and listing the
prerequisites for each of those targets, make can determine if they need to be
regenerated, and will handle doing so in the correct order, and where
possible, in parallel (if configured).