tweak implementation of canAddPatientTag, add tests
The implementation used to return (bool, error), however it was checking bool,
and if it was false, then it was assuming there was an error. In the
implementation, this was true, but with that being the case, there's no need
to return both the bool and the error. One can signal "false" with the
existence of an error. This way there's no worry about there needing to be any
consistency between the boolean value and the presence of an error.

Tests were added too, since there weren't any, and to help illustrate that the
names of the tags must be duplicates to trigger an error. It strikes me as odd
that tags with the same name, but different IDs don't trigger the duplicate
error. It's possible that's a bug, but I didn't want to change the
implementation without knowing better. Hopefully the test makes it clear what
the actual behavior is, even if it's not the intended behavior.