batch patient lists for clinic merges
Our first real attempt to generate a clinic merge report failed
because the patient list exceeded MongoDB's 16 MB Document limit.

Some quick math indicated that more than 1,600 patients could trigger
that limit, and so this code modifies the merge process to list
patients in batches of 1,024 at a time.

The test for this is very slow (due to patient creation), so I welcome
suggestions for improving that (with the understanding that they might
have to come later, this is a high priority backend hero ticket at the
moment.)

SIR-603