instead of updating patient sites when renamed, manually update the patients
During a clinic merge, there exists a case where a patient that is
updated, then deleted, in the same transaction, can cause a CDC event
with an empty document, which crashes the CDC.

When merging two clinics, imagine if the source clinic has a patient
that has the site A assigned. This patient will be merged into an
existing patient in the target clinic, meaning their source clinic
patient record will be deleted.

If the target clinic already has a site A, then the site from the
source clinic will be renamed (probably to "A (2)").

When the sites are merged, all patients in the source clinic with site
A will be updated to rename the site, but then, because that patient
is being merged into another existing patient in the target clinic, it
is then deleted. It's this process of an update, followed by a delete
that we believe is causing the issue.

BACK-3632