use the -o flag of oapi-codegen to prevent warnings in code
For some reason, oapi-codegen emits warnings to stdout, instead of
stderr. This means that when running the tool and redirecting the
output to a file, those warnings are captured in the generated source
file. To work around this, we were doing some sed processing, which
proved not to be platform-indepedent.

But oapi-codegen has a -o flag (maybe it didn't before? not sure) that
can be used to sidestep the entire situation. So... let's just do that
instead.